requirements.txt
streamlit==1.27.2
tensorflow-cpu==2.13.0
pillow
numpy
gdown
